requirements.txt
discord.py>=2.3.0
yt-dlp>=2023.7.6
PyNaCl>=1.5.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
psutil>=5.9.0